Undo merge
Undoing merge as merge contains changes associated with casacore 3.4.0

+++ ms/MSOper/MSSummary.h
@@ -110,7 +110,7 @@ public:
   ~MSSummary();
 
 // Retrieve number of rows
-   Int64 nrow() const;
+   Int nrow() const;
  
 // Retrieve image name
    String name() const;
